Split eip7594 into fk20 & recovery files
