+++ src/common/utils.h
@@ -34,6 +34,7 @@ extern "C" {
 bool is_power_of_two(uint64_t n);
 int log2_pow2(uint32_t n);
 uint32_t reverse_bits(uint32_t n);
+uint32_t reverse_bits_limited(uint32_t n, uint32_t value);
 C_KZG_RET bit_reversal_permutation(void *values, size_t size, uint64_t n);
 void compute_powers(fr_t *out, const fr_t *x, uint64_t n);
 bool pairings_verify(const g1_t *a1, const g2_t *a2, const g1_t *b1, const g2_t *b2);
