+++ src/eip7594/eip7594.c
@@ -20,7 +20,9 @@
 #include "common/lincomb.h"
 #include "common/utils.h"
 #include "eip7594/fft.h"
+#include "eip7594/fk20.h"
 #include "eip7594/poly.h"
+#include "eip7594/recovery.h"
 
 #include <assert.h> /* For assert */
 #include <string.h> /* For memcpy */
@@ -38,510 +40,6 @@
 
 /** The domain separator for verify_cell_kzg_proof_batch's random challenge. */
 static const char *RANDOM_CHALLENGE_DOMAIN_VERIFY_CELL_KZG_PROOF_BATCH = "RCKZGCBATCH__V1_";
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-// Helper Functions
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * Reverse the low-order bits in a 32-bit integer.
- *
- * @param[in]   n       To reverse `b` bits, set `n = 2 ^ b`
- * @param[in]   value   The bits to be reversed
- *
- * @return The reversal of the lowest log_2(n) bits of the input value.
- *
- * @remark n must be a power of two.
- */
-static uint32_t reverse_bits_limited(uint32_t n, uint32_t value) {
-    size_t unused_bit_len = 32 - log2_pow2(n);
-    return reverse_bits(value) >> unused_bit_len;
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-// Vanishing Polynomial
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * Calculates the minimal polynomial that evaluates to zero for each root.
- *
- * Uses straightforward long multiplication to calculate the product of `(x - r_i)` where `r_i` is
- * the i'th root. This results in a poly of degree roots_len.
- *
- * @param[in,out]   poly         The zero polynomial for roots
- * @param[in,out]   poly_len     The length of poly
- * @param[in]       roots        The array of roots
- * @param[in]       roots_len    The number of roots
- * @param[in]       s            The trusted setup
- *
- * @remark These do not have to be roots of unity. They are roots of a polynomial.
- * @remark `poly_len` must be at least `roots_len + 1` in length.
- */
-static C_KZG_RET compute_vanishing_polynomial_from_roots(
-    fr_t *poly, size_t *poly_len, const fr_t *roots, size_t roots_len
-) {
-    fr_t neg_root;
-
-    if (roots_len == 0) {
-        return C_KZG_BADARGS;
-    }
-
-    /* Initialize with -root[0] */
-    blst_fr_cneg(&poly[0], &roots[0], true);
-
-    for (size_t i = 1; i < roots_len; i++) {
-        blst_fr_cneg(&neg_root, &roots[i], true);
-
-        poly[i] = neg_root;
-        blst_fr_add(&poly[i], &poly[i], &poly[i - 1]);
-
-        for (size_t j = i - 1; j > 0; j--) {
-            blst_fr_mul(&poly[j], &poly[j], &neg_root);
-            blst_fr_add(&poly[j], &poly[j], &poly[j - 1]);
-        }
-        blst_fr_mul(&poly[0], &poly[0], &neg_root);
-    }
-
-    poly[roots_len] = FR_ONE;
-    *poly_len = roots_len + 1;
-
-    return C_KZG_OK;
-}
-
-/**
- * Computes the minimal polynomial that evaluates to zero at equally spaced chosen roots of unity in
- * the domain of size `FIELD_ELEMENTS_PER_BLOB`.
- *
- * The roots of unity are chosen based on the missing cell indices. If the i'th cell is missing,
- * then the i'th root of unity from `roots_of_unity` will be zero on the polynomial
- * computed, along with every `CELLS_PER_EXT_BLOB` spaced root of unity in the domain.
- *
- * @param[in,out]   vanishing_poly          The vanishing polynomial
- * @param[in]       missing_cell_indices    The array of missing cell indices
- * @param[in]       len_missing_cells       The number of missing cell indices
- * @param[in]       s                       The trusted setup
- *
- * @remark When all of the cells are missing, this algorithm has an edge case. We return
- * C_KZG_BADARGS in that case.
- * @remark When none of the cells are missing, recovery is trivial. We expect the caller to handle
- * this case, and return C_KZG_BADARGS if not.
- * @remark `missing_cell_indices` are assumed to be less than `CELLS_PER_EXT_BLOB`.
- */
-static C_KZG_RET vanishing_polynomial_for_missing_cells(
-    fr_t *vanishing_poly,
-    const uint64_t *missing_cell_indices,
-    size_t len_missing_cells,
-    const KZGSettings *s
-) {
-    C_KZG_RET ret;
-    fr_t *roots = NULL;
-    fr_t *short_vanishing_poly = NULL;
-    size_t short_vanishing_poly_len = 0;
-
-    /* Return early if none or all of the cells are missing */
-    if (len_missing_cells == 0 || len_missing_cells == CELLS_PER_EXT_BLOB) {
-        ret = C_KZG_BADARGS;
-        goto out;
-    }
-
-    /* Allocate arrays */
-    ret = new_fr_array(&roots, len_missing_cells);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&short_vanishing_poly, (len_missing_cells + 1));
-    if (ret != C_KZG_OK) goto out;
-
-    /*
-     * For each missing cell index, choose the corresponding root of unity from the subgroup of
-     * size `CELLS_PER_EXT_BLOB`.
-     *
-     * In other words, if the missing index is `i`, then we add \omega^i to the roots array, where
-     * \omega is a primitive `CELLS_PER_EXT_BLOB` root of unity.
-     */
-    size_t stride = FIELD_ELEMENTS_PER_EXT_BLOB / CELLS_PER_EXT_BLOB;
-    for (size_t i = 0; i < len_missing_cells; i++) {
-        roots[i] = s->roots_of_unity[missing_cell_indices[i] * stride];
-    }
-
-    /* Compute the polynomial that evaluates to zero on the roots */
-    ret = compute_vanishing_polynomial_from_roots(
-        short_vanishing_poly, &short_vanishing_poly_len, roots, len_missing_cells
-    );
-    if (ret != C_KZG_OK) goto out;
-
-    /*
-     * For each root \omega^i in `short_vanishing_poly`, we compute a polynomial that has roots at
-     *
-     *  H = {
-     *      \omega^i * \gamma^0,
-     *      \omega^i * \gamma^1,
-     *      ...,
-     *      \omega^i * \gamma^{FIELD_ELEMENTS_PER_CELL-1}
-     *  }
-     *
-     * where \gamma is a primitive `FIELD_ELEMENTS_PER_EXT_BLOB`-th root of unity.
-     *
-     * This is done by shifting the degree of all coefficients in `short_vanishing_poly` up by
-     * `FIELD_ELEMENTS_PER_CELL` amount.
-     */
-    for (size_t i = 0; i < short_vanishing_poly_len; i++) {
-        vanishing_poly[i * FIELD_ELEMENTS_PER_CELL] = short_vanishing_poly[i];
-    }
-
-out:
-    c_kzg_free(roots);
-    c_kzg_free(short_vanishing_poly);
-    return ret;
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-// Cell Recovery
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * Helper function to check if a uint64 value is in an array.
- *
- * @param[in]   arr         The array
- * @param[in]   arr_size    The size of the array
- * @param[in]   value       The value we want to search
- *
- * @return True if the value is in the array, otherwise false.
- */
-static bool is_in_array(const uint64_t *arr, size_t arr_size, uint64_t value) {
-    for (size_t i = 0; i < arr_size; i++) {
-        if (arr[i] == value) {
-            return true;
-        }
-    }
-    return false;
-}
-
-/**
- * Given a dataset with up to half the entries missing, return the reconstructed original. Assumes
- * that the inverse FFT of the original data has the upper half of its values equal to zero.
- *
- * @param[out]  reconstructed_data_out   Preallocated array for recovered cells
- * @param[in]   cell_indices             The cell indices you have
- * @param[in]   num_cells                The number of cells that you have
- * @param[in]   cells                    The cells that you have
- * @param[in]   s                        The trusted setup
- *
- * @remark `recovered` and `cells` can point to the same memory.
- * @remark The array of cells must be 2n length and in the correct order.
- * @remark Missing cells should be equal to FR_NULL.
- */
-static C_KZG_RET recover_cells_impl(
-    fr_t *reconstructed_data_out,
-    const uint64_t *cell_indices,
-    size_t num_cells,
-    fr_t *cells,
-    const KZGSettings *s
-) {
-    C_KZG_RET ret;
-    uint64_t *missing_cell_indices = NULL;
-    fr_t *vanishing_poly_eval = NULL;
-    fr_t *vanishing_poly_coeff = NULL;
-    fr_t *extended_evaluation_times_zero = NULL;
-    fr_t *extended_evaluation_times_zero_coeffs = NULL;
-    fr_t *extended_evaluations_over_coset = NULL;
-    fr_t *vanishing_poly_over_coset = NULL;
-    fr_t *reconstructed_poly_coeff = NULL;
-    fr_t *cells_brp = NULL;
-
-    /* Allocate space for arrays */
-    ret = c_kzg_calloc(
-        (void **)&missing_cell_indices, FIELD_ELEMENTS_PER_EXT_BLOB, sizeof(uint64_t)
-    );
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&vanishing_poly_eval, FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&vanishing_poly_coeff, FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&extended_evaluation_times_zero, FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&extended_evaluation_times_zero_coeffs, FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&extended_evaluations_over_coset, FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&vanishing_poly_over_coset, FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&reconstructed_poly_coeff, FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&cells_brp, FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-
-    /* Bit-reverse the data points, stored in new array */
-    memcpy(cells_brp, cells, FIELD_ELEMENTS_PER_EXT_BLOB * sizeof(fr_t));
-    ret = bit_reversal_permutation(cells_brp, sizeof(fr_t), FIELD_ELEMENTS_PER_EXT_BLOB);
-    if (ret != C_KZG_OK) goto out;
-
-    /* Identify missing cells */
-    size_t len_missing = 0;
-    for (size_t i = 0; i < CELLS_PER_EXT_BLOB; i++) {
-        /* Iterate over each cell index and check if we have received it */
-        if (!is_in_array(cell_indices, num_cells, i)) {
-            /*
-             * If the cell is missing, bit reverse the index and add it to the
-             * missing array.
-             */
-            uint32_t brp_i = reverse_bits_limited(CELLS_PER_EXT_BLOB, i);
-            missing_cell_indices[len_missing++] = brp_i;
-        }
-    }
-
-    /* Check that we have enough cells */
-    assert(len_missing <= CELLS_PER_EXT_BLOB / 2);
-
-    /* Compute Z(x) in monomial form */
-    ret = vanishing_polynomial_for_missing_cells(
-        vanishing_poly_coeff, missing_cell_indices, len_missing, s
-    );
-    if (ret != C_KZG_OK) goto out;
-
-    /* Convert Z(x) to evaluation form */
-    ret = fr_fft(vanishing_poly_eval, vanishing_poly_coeff, FIELD_ELEMENTS_PER_EXT_BLOB, s);
-    if (ret != C_KZG_OK) goto out;
-
-    /* Compute (E*Z)(x) = E(x) * Z(x) in evaluation form over the FFT domain */
-    for (size_t i = 0; i < FIELD_ELEMENTS_PER_EXT_BLOB; i++) {
-        if (fr_is_null(&cells_brp[i])) {
-            extended_evaluation_times_zero[i] = FR_ZERO;
-        } else {
-            blst_fr_mul(&extended_evaluation_times_zero[i], &cells_brp[i], &vanishing_poly_eval[i]);
-        }
-    }
-
-    /* Convert (E*Z)(x) to monomial form  */
-    ret = fr_ifft(
-        extended_evaluation_times_zero_coeffs,
-        extended_evaluation_times_zero,
-        FIELD_ELEMENTS_PER_EXT_BLOB,
-        s
-    );
-    if (ret != C_KZG_OK) goto out;
-
-    /*
-     * Polynomial division by convolution: Q3 = Q1 / Q2 where
-     *   Q1 = (D * Z_r,I)(k * x)
-     *   Q2 = Z_r,I(k * x)
-     *   Q3 = D(k * x)
-     */
-    ret = coset_fft(
-        extended_evaluations_over_coset,
-        extended_evaluation_times_zero_coeffs,
-        FIELD_ELEMENTS_PER_EXT_BLOB,
-        s
-    );
-    if (ret != C_KZG_OK) goto out;
-
-    ret = coset_fft(
-        vanishing_poly_over_coset, vanishing_poly_coeff, FIELD_ELEMENTS_PER_EXT_BLOB, s
-    );
-    if (ret != C_KZG_OK) goto out;
-
-    /* The result of the division is Q3 */
-    for (size_t i = 0; i < FIELD_ELEMENTS_PER_EXT_BLOB; i++) {
-        fr_div(
-            &extended_evaluations_over_coset[i],
-            &extended_evaluations_over_coset[i],
-            &vanishing_poly_over_coset[i]
-        );
-    }
-
-    /*
-     * Note: After the above polynomial division, extended_evaluations_over_coset is the same
-     * polynomial as reconstructed_poly_over_coset in the spec.
-     */
-
-    /* Convert the evaluations back to coefficents */
-    ret = coset_ifft(
-        reconstructed_poly_coeff, extended_evaluations_over_coset, FIELD_ELEMENTS_PER_EXT_BLOB, s
-    );
-    if (ret != C_KZG_OK) goto out;
-
-    /*
-     * After unscaling the reconstructed polynomial, we have D(x) which evaluates to our original
-     * data at the roots of unity. Next, we evaluate the polynomial to get the original data.
-     */
-    ret = fr_fft(reconstructed_data_out, reconstructed_poly_coeff, FIELD_ELEMENTS_PER_EXT_BLOB, s);
-    if (ret != C_KZG_OK) goto out;
-
-    /* Bit-reverse the recovered data points */
-    ret = bit_reversal_permutation(
-        reconstructed_data_out, sizeof(fr_t), FIELD_ELEMENTS_PER_EXT_BLOB
-    );
-    if (ret != C_KZG_OK) goto out;
-
-out:
-    c_kzg_free(missing_cell_indices);
-    c_kzg_free(vanishing_poly_eval);
-    c_kzg_free(extended_evaluation_times_zero);
-    c_kzg_free(extended_evaluation_times_zero_coeffs);
-    c_kzg_free(extended_evaluations_over_coset);
-    c_kzg_free(vanishing_poly_over_coset);
-    c_kzg_free(reconstructed_poly_coeff);
-    c_kzg_free(vanishing_poly_coeff);
-    c_kzg_free(cells_brp);
-    return ret;
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-// Cell Proofs
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * Reorder and extend polynomial coefficients for the toeplitz method, strided version.
- *
- * @param[out]  out     The reordered polynomial, size `n * 2 / stride`
- * @param[in]   in      The input polynomial, size `n`
- * @param[in]   n       The size of the input polynomial
- * @param[in]   offset  The offset
- * @param[in]   stride  The stride
- */
-static C_KZG_RET toeplitz_coeffs_stride(
-    fr_t *out, const fr_t *in, size_t n, uint64_t offset, uint64_t stride
-) {
-    uint64_t k, k2;
-
-    if (stride == 0) return C_KZG_BADARGS;
-
-    k = n / stride;
-    k2 = k * 2;
-
-    out[0] = in[n - 1 - offset];
-    for (uint64_t i = 1; i <= k + 1 && i < k2; i++) {
-        out[i] = FR_ZERO;
-    }
-    for (uint64_t i = k + 2, j = 2 * stride - offset - 1; i < k2; i++, j += stride) {
-        out[i] = in[j];
-    }
-
-    return C_KZG_OK;
-}
-
-/**
- * Compute FK20 cell-proofs for a polynomial.
- *
- * @param[out]  out An array of CELLS_PER_EXT_BLOB proofs
- * @param[in]   p   The polynomial, an array of coefficients
- * @param[in]   n   The length of the polynomial
- * @param[in]   s   The trusted setup
- *
- * @remark The polynomial should have FIELD_ELEMENTS_PER_BLOB coefficients. Only the lower half of
- * the extended polynomial is supplied because the upper half is assumed to be zero.
- */
-static C_KZG_RET compute_fk20_proofs(g1_t *out, const fr_t *p, size_t n, const KZGSettings *s) {
-    C_KZG_RET ret;
-    uint64_t k, k2;
-
-    blst_scalar *scalars = NULL;
-    fr_t **coeffs = NULL;
-    fr_t *toeplitz_coeffs = NULL;
-    fr_t *toeplitz_coeffs_fft = NULL;
-    g1_t *h = NULL;
-    g1_t *h_ext_fft = NULL;
-    void *scratch = NULL;
-    bool precompute = s->wbits != 0;
-
-    /* Initialize length variables */
-    k = n / FIELD_ELEMENTS_PER_CELL;
-    k2 = k * 2;
-
-    /* Do allocations */
-    ret = new_fr_array(&toeplitz_coeffs, k2);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_fr_array(&toeplitz_coeffs_fft, k2);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_g1_array(&h_ext_fft, k2);
-    if (ret != C_KZG_OK) goto out;
-    ret = new_g1_array(&h, k2);
-    if (ret != C_KZG_OK) goto out;
-
-    if (precompute) {
-        /* Allocations for fixed-base MSM */
-        ret = c_kzg_malloc(&scratch, s->scratch_size);
-        if (ret != C_KZG_OK) goto out;
-        ret = c_kzg_calloc((void **)&scalars, FIELD_ELEMENTS_PER_CELL, sizeof(blst_scalar));
-        if (ret != C_KZG_OK) goto out;
-    }
-
-    /* Allocate 2d array for coefficients by column */
-    ret = c_kzg_calloc((void **)&coeffs, k2, sizeof(void *));
-    if (ret != C_KZG_OK) goto out;
-    for (uint64_t i = 0; i < k2; i++) {
-        ret = new_fr_array(&coeffs[i], k);
-        if (ret != C_KZG_OK) goto out;
-    }
-
-    /* Initialize values to zero */
-    for (uint64_t i = 0; i < k2; i++) {
-        h_ext_fft[i] = G1_IDENTITY;
-    }
-
-    /* Compute toeplitz coefficients and organize by column */
-    for (uint64_t i = 0; i < FIELD_ELEMENTS_PER_CELL; i++) {
-        ret = toeplitz_coeffs_stride(toeplitz_coeffs, p, n, i, FIELD_ELEMENTS_PER_CELL);
-        if (ret != C_KZG_OK) goto out;
-        ret = fr_fft(toeplitz_coeffs_fft, toeplitz_coeffs, k2, s);
-        if (ret != C_KZG_OK) goto out;
-        for (uint64_t j = 0; j < k2; j++) {
-            coeffs[j][i] = toeplitz_coeffs_fft[j];
-        }
-    }
-
-    /* Compute h_ext_fft via MSM */
-    for (uint64_t i = 0; i < k2; i++) {
-        if (precompute) {
-            /* Transform the field elements to 255-bit scalars */
-            for (uint64_t j = 0; j < FIELD_ELEMENTS_PER_CELL; j++) {
-                blst_scalar_from_fr(&scalars[j], &coeffs[i][j]);
-            }
-            const byte *scalars_arg[2] = {(byte *)scalars, NULL};
-
-            /* A fixed-base MSM with precomputation */
-            blst_p1s_mult_wbits(
-                &h_ext_fft[i],
-                s->tables[i],
-                s->wbits,
-                FIELD_ELEMENTS_PER_CELL,
-                scalars_arg,
-                BITS_PER_FIELD_ELEMENT,
-                scratch
-            );
-        } else {
-            /* A pretty fast MSM without precomputation */
-            ret = g1_lincomb_fast(
-                &h_ext_fft[i], s->x_ext_fft_columns[i], coeffs[i], FIELD_ELEMENTS_PER_CELL
-            );
-            if (ret != C_KZG_OK) goto out;
-        }
-    }
-
-    ret = g1_ifft(h, h_ext_fft, k2, s);
-    if (ret != C_KZG_OK) goto out;
-
-    /* Zero the second half of h */
-    for (uint64_t i = k; i < k2; i++) {
-        h[i] = G1_IDENTITY;
-    }
-
-    ret = g1_fft(out, h, k2, s);
-    if (ret != C_KZG_OK) goto out;
-
-out:
-    c_kzg_free(scalars);
-    if (coeffs != NULL) {
-        for (uint64_t i = 0; i < k2; i++) {
-            c_kzg_free(coeffs[i]);
-        }
-        c_kzg_free(coeffs);
-    }
-    c_kzg_free(toeplitz_coeffs);
-    c_kzg_free(toeplitz_coeffs_fft);
-    c_kzg_free(h);
-    c_kzg_free(h_ext_fft);
-    c_kzg_free(scratch);
-    return ret;
-}
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 // Helper Functions for Batch Cell Verification
@@ -903,9 +401,7 @@ C_KZG_RET recover_cells_and_kzg_proofs(
         memcpy(recovered_cells, cells, CELLS_PER_EXT_BLOB * sizeof(Cell));
     } else {
         /* Perform cell recovery */
-        ret = recover_cells_impl(
-            recovered_cells_fr, cell_indices, num_cells, recovered_cells_fr, s
-        );
+        ret = recover_cells(recovered_cells_fr, cell_indices, num_cells, recovered_cells_fr, s);
         if (ret != C_KZG_OK) goto out;
 
         /* Convert the recovered data points to byte-form */
